% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_potential.R
\name{model_yield_potential}
\alias{model_yield_potential}
\title{Model yield potential of a crop.}
\usage{
model_yield_potential(
  date,
  par,
  tmin,
  tmax,
  params = list(x = 0.006, max_lai = 5, k = 0.68, tbase = 6, t_threshold = -2,
    spring_months = 1:5, fall_months = 8:12, rue = 1.91)
)
}
\arguments{
\item{date}{A vector of dates. Must be the same length as \code{tmin}, \code{tmax}, and \code{par}.}

\item{par}{A vector of daily photosynthetically active radiation (PAR) values in \eqn{\mathrm{MJ \cdot m^{-2} \cdot day^{-1}}}.}

\item{tmin}{A vector of daily minimum temperature values in \eqn{^\circ \mathrm{C}}.}

\item{tmax}{A vector of daily maximum temperature values in \eqn{^\circ \mathrm{C}}.}

\item{params}{A list of parameters that control the model. The list must contain the following elements:
\itemize{
\item{x}{ - The slope of the LAI curve. Default is 0.006.}
\item{max_lai}{ - The maximum LAI value. Default is 5}
\item{k}{ - The extinction coefficient. Default is 0.68.}
\item{tbase}{ - The base temperature for calculating GDD. Default is 6 \eqn{^\circ \mathrm{C}}.}
\item{t_threshold}{ - The temperature threshold for determining the start and end of the growing season. Default is -2 \eqn{^\circ \mathrm{C}}.}
\item{spring_months}{ - A vector of integers representing the months of the year that are considered spring. Default is 1:5.}
\item{fall_months}{ - A vector of integers representing the months of the year that are considered fall. Default is 8:12.}
\item{rue}{ - The radiation use efficiency. Default is 1.91 \eqn{\mathrm{g \cdot MJ^{-1}}}.}
}}
}
\description{
The \code{model_yield_potential()} calculates the yield potential of a crop
given a full year's worth of daily weather data.
}
